+++ styles/home.ts
@@ -52,6 +52,7 @@ export const SectionTitle = styled.h2`
   padding: 26px 36px 24px;
   color: var(--section-text-color);
   position: relative;
+  font-size: 20px;
   &::after {
     content: '';
     display: block;
@@ -147,7 +148,7 @@ export const NewContentBox = styled.div`
   transform: translate(-50%, 55%); */
   figure {
     width: 100%;
-    height: 78%;
+    height: 75%;
     /* margin: 0 auto; */
     img {
       width: 100%;
@@ -159,6 +160,7 @@ export const NewContentBox = styled.div`
   @media screen and (min-width: 768px) {
     padding: 0 12px 80px;
     bottom: 50%;
+    top: auto;
     transform: translate(-50%, 50%);
     figure {
       height: 86%;
@@ -205,19 +207,21 @@ export const MainCardInfo = styled.div`
 export const MainBannerBox = styled.div<mainBannerProps>`
   max-width: 768px;
   width: 100%;
-  padding: 16px 24px;
+  padding: 0px 24px;
   display: flex;
   justify-content: space-between;
   gap: 46px;
   align-items: center;
   position: absolute;
-  bottom: 90px;
+  top: 70px;
   left: 50%;
   transform: translateX(-50%);
   transition: all 0.3s ease-in-out;
   color: #fff;
 
   @media screen and (min-width: 768px) {
+    bottom: 90px;
+    top: auto;
     padding: 0;
     justify-content: flex-start;
     gap: 16px;
@@ -227,13 +231,13 @@ export const MainBannerBox = styled.div<mainBannerProps>`
 export const RandomText = styled.p<mainBannerProps>`
   width: 100%;
   display: flex;
-  padding: 8px 24px;
+  padding: 8px 16px;
   border-radius: 30px;
   flex-direction: column;
   word-wrap: break-word;
   font-size: 14px;
   font-weight: 700;
-  line-height: 28px;
+  line-height: 1.4;
   text-align: center;
   background-color: var(--random-text-color);
   color: #fff;
